Refactor: Move checkInputsForDepositTxRequest to InputsForDepositTxRequestValidation
