+++ core/src/main/java/bisq/core/trade/TradeManager.java
@@ -133,7 +133,7 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 
-import static bisq.core.trade.validation.DepositTxValidation.checkInputsForDepositTxRequest;
+import static bisq.core.trade.validation.InputsForDepositTxRequestValidation.checkInputsForDepositTxRequest;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
